renamed DayPanels to ActivityPanels
