+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/ArduinoLogo.jsx
