+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/BlocklyCanvasPanel.jsx
