+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/Icons/ArduinoLogo.jsx
