+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/Icons/PlotterLogo.jsx
