+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/Icons/textIcon.json
