+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/canvas/ContentCreatorCanvas.jsx
