+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/canvas/PublicCanvas.jsx
