+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/canvas/StudentCanvas.jsx
