+++ client/src/components/ActivityPanels/BlocklyCanvasPanel/modals/DisplayDiagramModal.jsx
