+++ client/src/components/ActivityPanels/DayPanels.less
