+++ client/src/components/ActivityPanels/Utils/helpers.js
