+++ client/src/views/Workspace/Workspace.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { getDayToolbox } from '../../Utils/requests.js';
-import BlocklyCanvasPanel from '../../components/DayPanels/BlocklyCanvasPanel/BlocklyCanvasPanel';
+import BlocklyCanvasPanel from '../../components/ActivityPanels/BlocklyCanvasPanel/BlocklyCanvasPanel';
 import { message } from 'antd';
 import NavBar from '../../components/NavBar/NavBar';
 import { useNavigate } from 'react-router-dom';
